test: bound the MCP pool suite's footprint in the parallel run (#563)
npm test runs one OS process per test *file*, all in parallel. The pool work
landed two new files, and this pair was the only place in the suite spawning
grandchild processes and booting loopback listeners, so it added the largest
per-file footprint of the branch to a run that is already CPU-saturated.

Fold the route-invalidation assertions into test/mcpPool.test.ts so the branch
adds one scheduled process instead of two; serve all three route assertions
from a single express listener instead of one per test (the store stubs are
stateless, so sharing is safe); and let AC4's "closeAll kills the child" ride
on AC1's already-pooled child instead of spawning its own. Both suites now
state concurrency: 1 explicitly — it is the default, but it is load-bearing
here because these tests count child processes.

Coverage is unchanged: same eight assertions, same acceptance criteria, all
still asserted on process identity and at the route boundary.
